Add keychain authentication support to verifier
